Contract Update

+++ app/Http/Controllers/Syndie/ContractObjective.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\Syndie;
 
+use App\Models\SyndieContractComment;
 use Illuminate\Http\Request;
 
 use App\Http\Requests;
@@ -21,11 +22,22 @@ class ContractObjective extends Controller
     public function getAdd(Request $request, $contract)
     {
         $contract = SyndieContract::findOrFail($contract);
+        if($request->user()->cannot('contract_moderate') && $contract->contractee_id != $request->user()->id)
+        {
+            abort('403','You do not have the required permission');
+        }
+
         return view('syndie.objective.add', ['contract' => $contract]);
     }
 
     public function postAdd(Request $request, $contract)
     {
+        $contract = SyndieContract::findOrFail($contract);
+        if($request->user()->cannot('contract_moderate') && $contract->contractee_id != $request->user()->id)
+        {
+            abort('403','You do not have the required permission');
+        }
+
         $this->validate($request, [
             'title' => 'required',
             'description' => 'required',
@@ -47,12 +59,24 @@ class ContractObjective extends Controller
     public function getEdit(Request $request, $objective)
     {
         $objective = SyndieContractObjective::findOrFail($objective);
+        $contract = $objective->contract()->get();
+        if($request->user()->cannot('contract_moderate') && $contract->contractee_id != $request->user()->id)
+        {
+            abort('403','You do not have the required permission');
+        }
+
         return view('syndie.objective.edit', ['objective' => $objective]);
     }
 
     public function postEdit(Request $request, $objective)
     {
         $objective = SyndieContractObjective::findOrFail($objective);
+        $contract = $objective->contract()->get();
+        if($request->user()->cannot('contract_moderate') && $contract->contractee_id != $request->user()->id)
+        {
+            abort('403','You do not have the required permission');
+        }
+
         $objective->title = $request->input('title');
         $objective->description = $request->input('description');
         $objective->reward_credits = $request->input('reward_credits');
@@ -65,6 +89,12 @@ class ContractObjective extends Controller
     public function close(Request $request, $objective)
     {
         $objective = SyndieContractObjective::findOrFail($objective);
+        $contract = $objective->contract()->get();
+        if($request->user()->cannot('contract_moderate') && $contract->contractee_id != $request->user()->id)
+        {
+            abort('403','You do not have the required permission');
+        }
+
         $objective->status = "closed";
         $objective->save();
 
@@ -74,6 +104,12 @@ class ContractObjective extends Controller
     public function open(Request $request, $objective)
     {
         $objective = SyndieContractObjective::findOrFail($objective);
+        $contract = $objective->contract()->get();
+        if($request->user()->cannot('contract_moderate') && $contract->contractee_id != $request->user()->id)
+        {
+            abort('403','You do not have the required permission');
+        }
+
         $objective->status = "open";
         $objective->save();
 
@@ -83,6 +119,12 @@ class ContractObjective extends Controller
     public function delete(Request $request, $objective)
     {
         $objective = SyndieContractObjective::findOrFail($objective);
+        $contract = $objective->contract()->get();
+        if($request->user()->cannot('contract_moderate') && $contract->contractee_id != $request->user()->id)
+        {
+            abort('403','You do not have the required permission');
+        }
+
         $objective->status = "deleted";
         $objective->save();
         $objective->delete();
